Separate User and Difficulty type from Matching Queue
