+++ matching-service/src/model/SelectedTopics.ts
@@ -1,0 +1,2 @@
+export type TDifficulty = "easy" | "medium" | "hard";
+export type SelectedDifficultyData = {[difficulty in TDifficulty] : boolean};
