+++ matching-service/src/model/User.ts
@@ -1,0 +1,11 @@
+import { SelectedDifficultyData } from "./SelectedTopics";
+
+type User = {
+    userToken : string,
+    difficulties : SelectedDifficultyData,
+    topics : string[]
+    isReady : boolean,
+    matchedUser : User | null
+}
+
+export { User };
